+++ src/conf/mod.rs
@@ -5,16 +5,6 @@ pub use parse::CmdOptConf;
 mod parse;
 
 impl CmdOptConf {
-    /*
-    pub fn base_dir(&self) -> String {
-        for o in self.opt_uc_x.iter() {
-            if let OptUcXParam::BaseDir(s) = o {
-                return s.clone();
-            }
-        }
-        String::new()
-    }
-    */
     pub fn is_opt_uc_x_help(&self) -> bool {
         for o in self.opt_uc_x.iter() {
             if let OptUcXParam::Help = o {