+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 * `execute_env()`
+* `base_dir=` of `-X` options
 
 ## [0.1.38] (2024-06-20)
 ### Added